aktualizr-lite: Add some testing for liteClient helper
This moves the liteClient function to helpers.{h|cc} and adds some
minimal testing to the function.
