+++ src/aktualizr_lite/CMakeLists.txt
@@ -19,10 +19,12 @@ add_test(test_aktualizr-lite
         ${RUN_VALGRIND}
 )
 add_library(t_lite-mock SHARED ostree_mock.cc)
+add_aktualizr_test(NAME lite-helpers SOURCES helpers.cc helpers_test.cc
+                   ARGS ${PROJECT_BINARY_DIR}/ostree_repo)
+set_tests_properties(test_lite-helpers PROPERTIES
+        ENVIRONMENT LD_PRELOAD=$<TARGET_FILE:t_lite-mock>)
 
 endif(BUILD_OSTREE)
-
-add_aktualizr_test(NAME lite-helpers SOURCES helpers.cc helpers_test.cc)
 
 aktualizr_source_file_checks(main.cc ${AKTUALIZR_LITE_SRC} ${AKTUALIZR_LITE_HEADERS} helpers_test.cc ostree_mock.cc)
 # vim: set tabstop=4 shiftwidth=4 expandtab:
