+++ src/aktualizr_lite/helpers.h
@@ -5,11 +5,15 @@
 
 #include <string.h>
 
+#include "primary/sotauptaneclient.h"
+
 struct Version {
   std::string raw_ver;
   Version(std::string version) : raw_ver(std::move(version)) {}
 
   bool operator<(const Version& other) { return strverscmp(raw_ver.c_str(), other.raw_ver.c_str()) < 0; }
 };
+
+std::shared_ptr<SotaUptaneClient> liteClient(Config& config, std::shared_ptr<INvStorage> storage);
 
 #endif  // AKTUALIZR_LITE_HELPERS
